feat: add AddPointForm component for new entry creation 🧹
Introduced a new `AddPointForm` component to streamline the process of adding new points to the application. This modularizes the form handling and includes date and time input, along with handling the submission of point data to the store. 

- Created `AddPointForm.svelte` to encapsulate form logic and UI.
- Integrated date and time inputs with a 'Now' button to preset current values.
- Axios submission now processes labels and points creation in a batch for efficiency.
- Replaced in-page logic in `+page.svelte` to use the new component, improving code clarity and maintainability.
- Adjusted various imports and method calls related to the restructuring of selectors for better organization. 
- Deleted outdated code related to the previous form handling logic to keep the project clean.
