+++ src/lib/components/LabelManager/StreamLabelManager.svelte
@@ -2,7 +2,7 @@
 import type { Store } from '#lib/core/replicache/store.js'
 import type { StreamId } from '#lib/ids.js'
 
-import { getLabelListGroupedByParent } from '#lib/core/select/label.js'
+import { getLabelListGroupedByParent } from '#lib/core/select/get-label-list-grouped-by-parent.js'
 
 import { query } from '#lib/utils/query.js'
